Modificando la estructura del proyecto
